Revert "drivers: tty: imx: fix flags of rs485 not work properly"
This reverts commit b8573c0ba6988f756f51ca5a9c90a19f2d7a379c.

This patch has been submitted upstream, but the tty/serial
maintainer raised some concerns about it, which were not
followed-up:

https://lore.kernel.org/linux-serial/7a31001d-da37-4deb-a366-084a4027bcd2@suse.cz/

To be on the safe side, let's revert it for now.

If this patch is reworked and accepted upstream, then it will be
applied via stable tree to 6.6 through the normal flow.

Let's not deviate from upstream unnecessarily here.

Signed-off-by: Fabio Estevam <festevam@gmail.com>